requirements.txt
flask
flask-openapi3-swagger
Flask-SQLAlchemy
SQLAlchemy
SQLAlchemy-Utils